SDK-71 move code to lib folder
