+++ lib/api-client.js
@@ -1,7 +1,7 @@
 const axios = require('axios');
 const get = require('lodash.get');
 const { StorageServerError, StorageClientError } = require('./errors');
-const pjson = require('./package.json');
+const pjson = require('../package.json');
 const { isValid, getErrorMessage } = require('./validation/utils');
 const { RecordResponseIO } = require('./validation/api-responses/record-response');
 const { FindResponseIO } = require('./validation/api-responses/find-response');
