+++ lib/in-crypt.js
