+++ lib/index.js
@@ -1,0 +1,7 @@
+const createStorage = require('./storage');
+const errors = require('./errors');
+
+module.exports = {
+  createStorage,
+  ...errors,
+};
