+++ lib/secret-key-accessor.js
