+++ lib/validation/api-responses/record-response.js
