+++ lib/validation/api-responses/write-response.js
