+++ lib/validation/custom-encryption-configs.js
