+++ lib/validation/record-key.js
