+++ lib/validation/secrets-data.js
