+++ lib/validation/storage-options.js
